requirements.txt
transformers==4.12.5
PyTorch==2.5.1
streamlit
scikit-learn==0.24.2
pandas